refactor(hooks): create `useBackendUrlBase` hook
