+++ src/components/occurrence-overview/Occurrence.tsx
@@ -5,10 +5,8 @@ import { useRouter } from 'next/router';
 import { useMemo, useState } from 'react';
 import Popup from 'reactjs-popup';
 import { GetOccurrencesByDateQuery } from 'src/graphql/graphql-types';
-import { GET_NAVIGATION, Navigation } from 'src/graphql/queries';
+import { useBackendUrlBase } from 'src/hooks';
 import { DATE_FORMAT, currentDate } from 'src/util';
-
-import { useQuery } from '@apollo/client';
 
 import styles from './Occurrence.module.scss';
 import OccurrenceComment from './comment';
@@ -47,16 +45,7 @@ const Occurrence = ({ occurrence }: Props) => {
     [routerLocale],
   );
 
-  // Get "URL-Base" from currently active backend
-  // Example: https://api.mensatt.de/v1/graphql => https://api.mensatt.de
-  const { data: navData } = useQuery<Navigation>(GET_NAVIGATION);
-  const backendURLBase = useMemo(() => {
-    const { protocol, hostname } = new URL(
-      navData?.backends[navData.activeBackendIdx].url ||
-        'https://api.mensatt.de/v1/graphql',
-    );
-    return `${protocol}//${hostname}`;
-  }, [navData]);
+  const backendUrlBase = useBackendUrlBase();
 
   const filteredDishReviews = useMemo(
     () =>
@@ -132,7 +121,7 @@ const Occurrence = ({ occurrence }: Props) => {
         {filteredDishImages.length > 0 ? (
           <Image
             key={randomImage.id}
-            src={backendURLBase + randomImage.imageUrl}
+            src={backendUrlBase + randomImage.imageUrl}
             alt={t('imageDescription', {
               name: occurrenceName,
               author: randomImage.displayName || t('noAuthorName'),
